Int Casting modified in Rectangle.cpp

+++ W8/lab/Rectangle.cpp
@@ -15,7 +15,7 @@ namespace sdds {
 
 	Rectangle::Rectangle(const char* label, int width, int height) 
 		: LblShape(label), m_width(width), m_height(height)  {
-		if (m_height < 3 || m_width < strlen(label) + 2)
+		if (m_height < 3 || m_width < (int)strlen(label) + 2)
 		{
 			m_width = 0;
 			m_height = 0;
@@ -31,7 +31,7 @@ namespace sdds {
 	}
 
 	void Rectangle::draw(ostream& os) const {
-		if (m_height >= 3 && m_width >= (strlen(label())) + 2)
+		if (m_height >= 3 && m_width >= (int)(strlen(label())) + 2)
 		{
 			os << '+' << std::string(m_width - 2, '-') << '+' << std::endl;
